Cleanups from stricter span usage of size_t, pt. 2: media
More use of first(), compile-time sizes, size_t, etc.

Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6045615
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1387694}

+++ media/formats/mp4/box_reader.h
@@ -6,6 +6,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef MEDIA_FORMATS_MP4_BOX_READER_H_
 #define MEDIA_FORMATS_MP4_BOX_READER_H_
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include <limits>
@@ -80,7 +81,7 @@ class MEDIA_EXPORT BufferReader {
 
   // Reads a sequence of bytes verbatim from the buffer into `t` after clearing
   // `t`, and advances the stream pointer.
-  [[nodiscard]] bool ReadVec(std::vector<uint8_t>* t, uint64_t count);
+  [[nodiscard]] bool ReadVec(std::vector<uint8_t>* t, size_t count);
 
   // Advance the stream by this many bytes.
   [[nodiscard]] bool SkipBytes(uint64_t nbytes);
