+++ media/muxers/mp4_muxer_delegate_unittest.cc
@@ -19,6 +19,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/containers/span_reader.h"
 #include "base/files/file_path.h"
 #include "base/files/memory_mapped_file.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/path_service.h"
 #include "base/test/bind.h"
 #include "base/test/task_environment.h"
@@ -1169,7 +1170,8 @@ TEST_F(Mp4MuxerDelegateTest, MfraBoxOnAudioAndVideoAddition) {
   reader.ReadU32BigEndian(value);
   EXPECT_EQ(value, 1u);  // sample number.
   EXPECT_EQ(base::span(total_written_data)
-                .subspan(moof_offset, third_moof_written_data.size()),
+                .subspan(base::checked_cast<size_t>(moof_offset),
+                         third_moof_written_data.size()),
             third_moof_written_data);
 
   // Fourth entry.
@@ -1188,7 +1190,8 @@ TEST_F(Mp4MuxerDelegateTest, MfraBoxOnAudioAndVideoAddition) {
   reader.ReadU32BigEndian(value);
   EXPECT_EQ(value, 1u);  // sample number.
   EXPECT_EQ(base::span(total_written_data)
-                .subspan(fourth_moof_offset, fourth_moof_written_data.size()),
+                .subspan(base::checked_cast<size_t>(fourth_moof_offset),
+                         fourth_moof_written_data.size()),
             fourth_moof_written_data);
 }
